Separate client from mongo custom formats
