+++ app/models/UserSummary.scala
@@ -3,9 +3,7 @@ package models
 import org.joda.time.DateTime
 import play.api.libs.json.Json
 import repositories.{IdentityUser, Orphan, PersistedUser}
-
 import scala.language.implicitConversions
-import MongoJsFormats._
 
 case class UserSummary(id: String,
                        email: String,
